+++ python/GafferSceneUI/PassesUI.py
@@ -1,0 +1,78 @@
+##########################################################################
+#
+#  Copyright (c) 2023, Cinesite VFX Ltd. All rights reserved.
+#
+#  Redistribution and use in source and binary forms, with or without
+#  modification, are permitted provided that the following conditions are
+#  met:
+#
+#      * Redistributions of source code must retain the above
+#        copyright notice, this list of conditions and the following
+#        disclaimer.
+#
+#      * Redistributions in binary form must reproduce the above
+#        copyright notice, this list of conditions and the following
+#        disclaimer in the documentation and/or other materials provided with
+#        the distribution.
+#
+#      * Neither the name of John Haddon nor the names of
+#        any other contributors to this software may be used to endorse or
+#        promote products derived from this software without specific prior
+#        written permission.
+#
+#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+#  IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+#  THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+#  PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+#  CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+#  EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+#  PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+#  PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+#  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+#  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+#  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
+##########################################################################
+
+import Gaffer
+import GafferScene
+
+Gaffer.Metadata.registerNode(
+
+	GafferScene.Passes,
+
+	"description",
+	"""
+	Appends passes to the scene globals.
+
+	Passes can be used to define named variations of a scene. These can
+	be rendered by dispatching a PassWedge node downstream of your render
+	node of choice, or written to disk by dispatching a PassWedge node
+	downstream of a SceneWriter.
+
+	Scenes can be varied per pass based on the value of the `pass`
+	context variable, which will contain the name of the current pass
+	being dispatched. `${pass}` can be used on the `selector` plug of
+	Spreadsheet or NameSwitch nodes to choose specific plug values or
+	branches of the node graph per pass, and its value can be queried
+	using Expression or ContextQuery nodes.
+
+	> Tip : If any of the specified names already exist, they will be
+	> removed from their existing position in the list and appended
+	> to the end.
+	""",
+
+	plugs = {
+
+		"names" : [
+
+			"description",
+			"""
+			The names of passes to be created.
+			""",
+
+		],
+
+	}
+
+)