+++ python/GafferSceneUI/__init__.py
@@ -187,6 +187,7 @@ from . import FramingConstraintUI
 from . import MeshNormalsUI
 from . import LightToolUI
 from . import ImageScatterUI
+from . import PassesUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.